Make getStakingData reusable
